refactor: Entity 파일 이동
